requirements.txt
anyio==3.6.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.79.1
h11==0.13.0
idna==3.3
Pillow==9.2.0
pydantic==1.9.2
requests==2.28.1
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.2
